+++ core/changelog.md
@@ -1,3 +1,4 @@
+- feat: added support for addWatermark parameter in Gemini Image Gen
 - fix: allow setting authorization header through extra headers and in allow list and deny list
 - feat: added image generation support for bedrock
 - feat: added support for gemini google search tool